more cleanup and tests
